+++ changelog.md
@@ -7,6 +7,7 @@
 * Fix formula bug caused when looking up date intervals.
 * Add debugging commands/options for inspecting tables and updating formulas.
 * Fix rare formula bug with multiple different formulas and view filters in one table.
+* Added FormulaField to the options for the primary field.
 
 ## Released (2022-10-05 1.10.0)